[Community page] add write, read page
